+++ app/user/store/entity/user.py
@@ -3,7 +3,7 @@ from uuid import uuid4
 
 from sqlalchemy.orm import mapped_column
 from sqlalchemy.orm.base import Mapped
-from user.model.entity.base import Base
+from user.store.entity.base import Base
 
 
 class User(Base):
